requirements.txt
django==2.1.7
markdown==3.0.1
djangorestframework==3.9.0
django-filter==2.0.0
django-webpack-loader==0.6.0
pillow==6.0.0